Allow mock pipes to be built outside of testing
In order to use mock pipes, we used to have to import pemi.testing, which required pytest.
This package may not and should not be available in a development environment.  However,
it can still be useful to mock a pipe while debugging a process (e.g., to make sure that
a target isn't accidentally loaded).  This commit allows one to mock pipes using pemi.pipe.mock_pipe.
